refactor(core): extract verification to verify.rs
Move verify function, ScriptVerificationFlags and ScriptVerifyError to
dedicated module.
